src/grna-modeling/requirements.txt
biopython>=1.79
biotite>=0.35.0
einops>=0.6.0
numpy<2.0.0
tokenizers
torch>=1.10
transformers>=4.16.2
flake8
black
isort
pre-commit
